+++ codegen/smithy-go-codegen/src/main/java/software/amazon/smithy/go/codegen/integration/ProtocolUtils.java
@@ -210,8 +210,4 @@ public final class ProtocolUtils {
 
         lambda.accept(acceptVar);
     }
-
-    public static void writeDocumentInterfaceCheck(GenerationContext context, GoWriter writer) {
-
-    }
 }
