+++ document/json/json.go
@@ -1,5 +1,6 @@
 package json
 
+// NewEncoder returns an Encoder for serializing Smithy documents for JSON based protocols.
 func NewEncoder(optFns ...func(options *EncoderOptions)) *Encoder {
 	o := EncoderOptions{}
 
@@ -12,6 +13,7 @@ func NewEncoder(optFns ...func(options *EncoderOptions)) *Encoder {
 	}
 }
 
+// NewDecoder returns a Decoder for deserializing Smithy documents for JSON based protocols.
 func NewDecoder(optFns ...func(*DecoderOptions)) *Decoder {
 	o := DecoderOptions{}
